+++ floogen/model/connection.py
@@ -5,12 +5,13 @@
 # Author: Tim Fischer <fischeti@iis.ee.ethz.ch>
 
 
-from pydantic import BaseModel, ConfigDict, field_validator, model_validator
+from pydantic import field_validator, model_validator
 
-from floogen.model.routing import XYDirections
+from floogen.model.config import ConfigModel, OneOrMany
+from floogen.model.routing import PortDirection
 
 
-class ConnectionDesc(BaseModel):
+class ConnectionDesc(ConfigModel):
     """Connection class to describe a connection between routers and endpoints.
 
     Attributes:
@@ -26,29 +27,19 @@ class ConnectionDesc(BaseModel):
         description (Optional[str]): Optional description of the connection.
     """
 
-    model_config = ConfigDict(extra="forbid")
-
     description: str | None = ""
     src: str
     dst: str
     src_range: list[tuple[int, int]] | None = None
     dst_range: list[tuple[int, int]] | None = None
-    src_idx: list[int] | None = None
-    dst_idx: list[int] | None = None
+    src_idx: OneOrMany[int] | None = None
+    dst_idx: OneOrMany[int] | None = None
     src_lvl: int | None = None
     dst_lvl: int | None = None
-    dst_dir: int | None = None
-    src_dir: int | None = None
+    dst_dir: PortDirection | None = None
+    src_dir: PortDirection | None = None
     allow_multi: bool | None = False
     bidirectional: bool | None = True
-
-    @field_validator("src_idx", "dst_idx", mode="before")
-    @classmethod
-    def int_to_list(cls, v):
-        """Convert int to list."""
-        if isinstance(v, int):
-            return [v]
-        return v
 
     @model_validator(mode="after")
     def check_indexing(self):
@@ -58,14 +49,6 @@ class ConnectionDesc(BaseModel):
         if self.dst_idx and self.dst_lvl:
             raise ValueError("dst_idx and dst_lvl are mutually exclusive")
         return self
-
-    @field_validator("src_dir", "dst_dir", mode="before")
-    @classmethod
-    def str_to_int(cls, v):
-        """Convert str to int."""
-        if isinstance(v, str):
-            return XYDirections[v.upper()].value
-        return v
 
     @field_validator("bidirectional", mode="after")
     @classmethod